requirements.txt
psycopg2-binary
pandas
sqlalchemy
boto3
snowflake-connector-python